paper-vera-rl/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-vera-rl"
version = "0.1.0"
description = "VERA-RL paper workspace"
requires-python = ">=3.10"
dependencies = ["agent-rl-core", "pyyaml>=6.0", "numpy>=1.26"]

[tool.setuptools.packages.find]
where = ["methods"]
